requirements.txt
Package         Version  
--------------- ---------
bcrypt          3.1.3    
boto3           1.4.7    
botocore        1.7.48   
bottle          0.12.13  
certifi         2019.9.11
cffi            1.13.2   
chardet         3.0.4    
Deprecated      1.2.3    
docutils        0.15.2   
gevent          1.2.2    
greenlet        0.4.15   
gunicorn        19.7.1   
idna            2.6      
jmespath        0.9.4    
marshmallow     2.13.6   
mongoengine     0.15.0   
numpy           1.17.4   
pandas          0.24.2   
peewee          2.10.1   
pip             19.3.1   
pycparser       2.19     
PyJWT           1.5.3    
pymongo         3.9.0    
PyMySQL         0.7.11   
python-dateutil 2.8.1    
pytz            2019.3   
redis           3.3.11   
requests        2.18.4   
s3transfer      0.1.13   
setuptools      41.0.1   
six             1.13.0   
urllib3         1.22     
voluptuous      0.10.5   
wheel           0.33.6   
wrapt           1.11.2   
xlrd            1.2.0    
